+++ applications/tabular/service/startTabular.js
@@ -1,0 +1,6 @@
+/**
+ * @fileOverview
+ * A script to launch the single Echo server.
+ */
+
+require('../lib/service').start(10093, 'alpha');